pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkml-openapi"
version = "0.8.2"
description = "Generate OpenAPI 3.1 specifications from LinkML schemas"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Jack Higgs"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "linkml-runtime>=1.7.0",
    "openapi-pydantic>=0.5.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "jinja2>=3.1",
]

[project.urls]
Homepage = "https://github.com/jackhiggs/linkml-openapi"
Repository = "https://github.com/jackhiggs/linkml-openapi"
Issues = "https://github.com/jackhiggs/linkml-openapi/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
    "openapi-spec-validator>=0.7.0",
]

[project.scripts]
gen-openapi = "linkml_openapi.cli:cli"
gen-spring-server = "linkml_openapi.spring.cli:cli"

[project.entry-points."linkml.generators"]
openapi = "linkml_openapi.generator:OpenAPIGenerator"
spring = "linkml_openapi.spring.generator:SpringServerGenerator"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring external tools (npx, docker)",
]
